pyproject.toml
[project]
name = "psychology-reports"
version = "0.1.0"
description = ""
package-mode = false
authors = [
    {name = "ranishot",email = "ranishot@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pwdlib (>=0.2.1,<0.3.0)",
    "minio (>=7.2.15,<8.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pymupdf (>=1.26.3,<2.0.0)",
    "pypdf (>=5.9.0,<6.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "img2pdf (>=0.6.1,<0.7.0)",
    "taskiq[redis] (>=0.11.18,<0.12.0)",
    "taskiq-redis (>=1.1.0,<2.0.0)",
    "taskiq-fastapi (>=0.3.5,<0.4.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
faker = "^37.3.0"

